requirements.txt
ae-ffmpeg==1.2.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
Authlib==1.3.1
auto-editor==26.0.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.138
botocore==1.34.138
Bottleneck==1.4.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
db-dtypes==1.3.0
decorator==4.4.2
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.6.1
gcsfs==2024.6.1
google==3.0.0
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud==0.34.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-documentai==2.31.0
google-cloud-documentai-toolbox==0.14.0a0
google-cloud-storage==2.18.2
google-cloud-vision==3.7.4
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
immutabledict==4.2.0
intervaltree==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
llvmlite==0.43.0
loadingspinner==1.0.1
lxml==5.3.0
Mako==1.3.5
MarkupSafe==3.0.1
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
moviepy==1.0.3
multidict==6.0.5
mysql-connector-python==9.0.0
numba==0.60.0
numexpr==2.10.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.43.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pandas-gbq==0.23.1
pikepdf==8.15.1
pillow==10.4.0
proglog==0.1.10
prompt-toolkit==3.0.39
proto-plus==1.24.0
protobuf==5.28.0
psutil==6.0.0
psycopg2-binary==2.9.9
pwinput==1.0.3
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyav==13.1.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydata-google-auth==1.8.2
PyJWT==2.8.0
pyparsing==3.1.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==308
rarfile==4.2
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.31
tabulate==0.9.0
termcolor==2.3.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
waitress==3.0.0
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.9.7
yaspin==3.1.0
